jail,vm: try to deletemacs only if they exist

+++ internal/services/jail/jail.go
@@ -778,27 +778,29 @@ func (s *Service) DeleteJail(ctId uint, deleteMacs bool) error {
 			}
 		}
 
-		tx := s.DB.Begin()
+		if len(usedMACS) > 0 {
+			tx := s.DB.Begin()
 
-		if err := tx.Where("object_id IN ?", usedMACS).
-			Delete(&networkModels.ObjectEntry{}).Error; err != nil {
-			tx.Rollback()
-			return fmt.Errorf("failed_to_delete_object_entries: %w", err)
-		}
+			if err := tx.Where("object_id IN ?", usedMACS).
+				Delete(&networkModels.ObjectEntry{}).Error; err != nil {
+				tx.Rollback()
+				return fmt.Errorf("failed_to_delete_object_entries: %w", err)
+			}
 
-		if err := tx.Where("object_id IN ?", usedMACS).
-			Delete(&networkModels.ObjectResolution{}).Error; err != nil {
-			tx.Rollback()
-			return fmt.Errorf("failed_to_delete_object_resolutions: %w", err)
-		}
+			if err := tx.Where("object_id IN ?", usedMACS).
+				Delete(&networkModels.ObjectResolution{}).Error; err != nil {
+				tx.Rollback()
+				return fmt.Errorf("failed_to_delete_object_resolutions: %w", err)
+			}
 
-		if err := tx.Delete(&networkModels.Object{}, usedMACS).Error; err != nil {
-			tx.Rollback()
-			return fmt.Errorf("failed_to_delete_objects: %w", err)
-		}
+			if err := tx.Delete(&networkModels.Object{}, usedMACS).Error; err != nil {
+				tx.Rollback()
+				return fmt.Errorf("failed_to_delete_objects: %w", err)
+			}
 
-		if err := tx.Commit().Error; err != nil {
-			return fmt.Errorf("failed_to_commit_cleanup: %w", err)
+			if err := tx.Commit().Error; err != nil {
+				return fmt.Errorf("failed_to_commit_cleanup: %w", err)
+			}
 		}
 	}
